Доработка функции конвертации

+++ src/external_api.py
@@ -9,7 +9,7 @@ api_key = os.getenv("API_KEY")  # pragma: no cover
 
 def convert_to_rubles(transaction: dict) -> float:
     """Функция принимает на вход транзакцию и возвращает сумму транзакции в рублях"""
-    amount = transaction["operationAmount"]["amount"]
+    amount = float(transaction["operationAmount"]["amount"])
     if (
         transaction["operationAmount"]["currency"]["code"] == "USD"
         or transaction["operationAmount"]["currency"]["code"] == "EUR"
@@ -19,6 +19,6 @@ def convert_to_rubles(transaction: dict) -> float:
 
         response = requests.get(url, headers=headers)
         data = response.json()
-        transfer_to_rubles = float(amount * data["rates"]["RUB"])
+        transfer_to_rubles = amount * data["rates"]["RUB"]
         return transfer_to_rubles
     return float(amount)
